Improve add-missing-to-queue with predicate filter
This allows passing a small python expression to filter
out any unwanted datasets before sending them to SQS.

Currently in DEA, our Water Observations datasets are only
created for inputs with a high geometric accuracy, BUT, this
filtering is implemented in the SQS Subscription and not in
Alchemist.

I'd like eventually to move this filtering capability into
the Alchemist configuration file to ensure it is always used
for certain products.